Added base_ingest_client.py for common functionality, and added ingest_from_stream to ingest_client
